Absorb memory editor into memories page as a modal
