--- app/memory-editor/page.js
@@ -1,5 +1,0 @@
-import MemoryEditor from "@/components/MemoryEditor";
-
-export default function MemoryEditorPage() {
-  return <MemoryEditor />;
-}
