--- components/MemoryEditor/MemoryEditor.module.scss
@@ -1,18 +1,0 @@
-.memoryEditor {
-  max-width: 600px;
-  margin: 0 auto;
-  position: relative;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 16px;
-}
-
-.name {
-  max-width: 400px;
-}
-
-.params {
-  max-width: 400px;
-}
