+++ components/MemoryEditorModal/MemoryEditorModal.module.scss
@@ -1,0 +1,11 @@
+.memoryEditor {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 0 16px 16px;
+  overflow: auto;
+}
+
+.params {
+  max-width: 400px;
+}
