+++ components/PIdolSelect/PIdolSelect.js
@@ -1,4 +1,4 @@
-import { memo, useContext, useMemo, useState } from "react";
+import { memo, useContext, useEffect, useMemo, useState } from "react";
 import { PIdols } from "gakumas-data";
 import PlanIdolSelects from "@/components/PlanIdolSelects";
 import PIdol from "@/components/PIdol";
@@ -8,6 +8,13 @@ import styles from "./PIdolSelect.module.scss";
 function PIdolSelect({ selected, onChange }) {
   const { plan, setPlan, idolId, setIdolId } = useContext(WorkspaceContext);
   const [expanded, setExpanded] = useState(!selected);
+
+  useEffect(() => {
+    if (!selected || !expanded) return;
+    const pIdol = PIdols.getById(selected);
+    setPlan(pIdol.plan);
+    setIdolId(pIdol.idolId);
+  }, [expanded, selected]);
 
   const pIdols = useMemo(
     () => PIdols.getFiltered({ plans: [plan], idolIds: [idolId] }),
